SaltTrackerService/SaltTrackerService/requirements.txt
Django==1.9.6
django-admin-sortable==2.1
django-modeltranslation==0.12
django-pyodbc-azure==1.9.6.0
djangorestframework==3.3.3
olefile==0.44
Pillow==4.1.1
pip==9.0.1
pyodbc==3.0.10
pytz==2016.4
requests==2.11.1
setuptools==15.0
unicodecsv==0.14.1
wheel==0.24.0
xlrd==1.0.0
XlsxWriter==0.9.3
